backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-CORS
Flask-Caching
Flask-APScheduler
python-dotenv
aiohttp
beautifulsoup4
selenium
nltk
mysqlclient